fix(warnings): clear libevmu build warnings, incl. three real defects
A clean six-slice `build_xcframework.py` run emitted 307 warning lines
(48 unique). Down to 64 lines / 10 unique, all of them upstream.

Three of the warnings were sitting on actual bugs, not narrowing noise:

- `evmu_lcd.c` passed `GblVariant_setEnum`/`setFlags` their arguments in
  the wrong order — the signature is `(variant, GblType, value)` and the
  code passed `(variant, value, GblType)`. Property reads of the LCD's
  `refreshRate` and `icons` therefore produced a variant built from a
  bogus type. `evmu_rom.c` already had the order right.
- `gyro_vmu_flash.c` tested a `FILE*` for openness with
  `int wasOpen = (uintptr_t)fp`, truncating the pointer to 32 bits. Any
  handle whose low word happened to be zero read back as "not open".
- The same file declared `bytesRead`/`bytesTotal`, zeroed them, and never
  assigned them, so the post-`fread` check compared the read size against
  a constant 0. Loading a perfectly good flash image always logged "All
  bytes were not read properly" and "Read 0 bytes". Uses fread()'s return
  value now.

Everything else was a genuine but harmless narrowing — VMU addresses are
16-bit and flash is 128 KB, so `size_t` → `EvmuAddress`/`int` can't lose
anything — made explicit at each site. `EvmuFat_toBlocks` was doing its
division through `div()`, which is int-only; it divides in `size_t` now.

Left alone: six `size_t` → `GblEnum` narrowings that come from upstream
libGimbal's `GBL_PROPERTIES_REGISTER` macro expanding in our files, and
libtool's "has no symbols" on four legitimately-empty translation units
(the merge step suppresses its copy; Xcode's per-target archive step
ignores `OTHER_LIBTOOLFLAGS`, verified).

+++ legacy/source/formats/gyro_vmu_vmi.c
@@ -84,7 +84,7 @@ void gyVmuVmiGenerateFromVms(VMIFileInfo* vmi, const VMSFileInfo* vms, size_t vm
     vmi->creationHour       = tm->tm_hour;
     vmi->creationSecond     = tm->tm_sec;
     vmi->creationWeekday    = tm->tm_wday;
-    vmi->fileSize           = /*gyVmuVmsFileInfoHeaderSize(vms) + */ vmsFileSize;
+    vmi->fileSize           = (uint32_t)(/*gyVmuVmsFileInfoHeaderSize(vms) + */ vmsFileSize);
     vmi->checksum           = gblHashCrc16BitPartial((const unsigned char *)vms, vmi->fileSize, NULL);
     gyVmuFlashPrintVMIFileInfo(vmi);
